Move resolver-related code into resolver module
