+++ frontend/src/metabase-lib/v1/expressions/resolver.ts
@@ -1,9 +1,10 @@
 import { c, t } from "ttag";
 
-import type * as Lib from "metabase-lib";
+import * as Lib from "metabase-lib";
 
+import { EDITOR_FK_SYMBOLS } from "./config";
 import { ResolverError } from "./errors";
-import { parseDimension, parseMetric, parseSegment } from "./identifier";
+import { getDisplayNameWithSeparator } from "./identifier";
 import type { Node } from "./pratt";
 
 export type Kind = "field" | "metric" | "segment";
@@ -56,4 +57,111 @@ export function resolver(options: {
       return dimension;
     }
   };
+}
+
+function parseMetric(
+  metricName: string,
+  { query, stageIndex }: { query: Lib.Query; stageIndex: number },
+) {
+  const metrics = Lib.availableMetrics(query, stageIndex);
+
+  const metric = metrics.find((metric) => {
+    const displayInfo = Lib.displayInfo(query, stageIndex, metric);
+
+    return displayInfo.displayName.toLowerCase() === metricName.toLowerCase();
+  });
+
+  if (metric) {
+    return metric;
+  }
+}
+
+function parseSegment(
+  segmentName: string,
+  {
+    query,
+    stageIndex,
+    expressionIndex,
+  }: { query: Lib.Query; stageIndex: number; expressionIndex?: number },
+) {
+  const segment = Lib.availableSegments(query, stageIndex).find((segment) => {
+    const displayInfo = Lib.displayInfo(query, stageIndex, segment);
+
+    return displayInfo.displayName.toLowerCase() === segmentName.toLowerCase();
+  });
+
+  if (segment) {
+    return segment;
+  }
+
+  const column = Lib.expressionableColumns(
+    query,
+    stageIndex,
+    expressionIndex,
+  ).find((field) => {
+    const displayInfo = Lib.displayInfo(query, stageIndex, field);
+    return displayInfo.displayName.toLowerCase() === segmentName.toLowerCase();
+  });
+
+  if (column && Lib.isBoolean(column)) {
+    return column;
+  }
+}
+
+export function parseDimension(
+  name: string,
+  options: {
+    query: Lib.Query;
+    stageIndex: number;
+    expressionIndex?: number | undefined;
+    startRule: string;
+  },
+) {
+  return getAvailableDimensions(options).find(({ info }) => {
+    return EDITOR_FK_SYMBOLS.symbols.some((separator) => {
+      const displayName = getDisplayNameWithSeparator(
+        info.longDisplayName,
+        separator,
+      );
+
+      return displayName === name;
+    });
+  })?.dimension;
+}
+
+function getAvailableDimensions({
+  query,
+  stageIndex,
+  expressionIndex,
+  startRule,
+}: {
+  query: Lib.Query;
+  stageIndex: number;
+  expressionIndex?: number | undefined;
+  startRule: string;
+}) {
+  const results = Lib.expressionableColumns(
+    query,
+    stageIndex,
+    expressionIndex,
+  ).map((dimension) => {
+    return {
+      dimension,
+      info: Lib.displayInfo(query, stageIndex, dimension),
+    };
+  });
+
+  if (startRule === "aggregation") {
+    return [
+      ...results,
+      ...Lib.availableMetrics(query, stageIndex).map((dimension) => {
+        return {
+          dimension,
+          info: Lib.displayInfo(query, stageIndex, dimension),
+        };
+      }),
+    ];
+  }
+
+  return results;
 }
